# JSON serialization performance improvement